Removes test inheritance and update suppresswarning note
