+++ sample-resource-plugin/src/main/java/org/opensearch/sample/resource/actions/rest/create/CreateResourceRestAction.java
@@ -92,7 +92,7 @@ public class CreateResourceRestAction extends BaseRestHandler {
         );
     }
 
-    // NOTE: Suppressing warnings should be avoided as it may lead to loosing important information for while root-causing an issue
+    // NOTE: Do NOT use @SuppressWarnings("unchecked") on untrusted data in production code. This is used here only to keep the code simple
     @SuppressWarnings("unchecked")
     private Map<String, String> getAttributes(Map<String, Object> source) {
         return source.containsKey("attributes") ? (Map<String, String>) source.get("attributes") : null;
